reviewed for #53 and brought up some ideas
There are some things I am not fully sure how we are implementing and we should discuss how it will work in person to understand easier. I brought up some issues but it might just be issues for the way that I see of doing things, not how you were thinking, i can revert them if needed later.